Introducing a first coarse version workflow
 - CityGMLFileFromArchive extracted from demo_lyon_metropole_dowload_and_sanitize.py
   and placed in city_gml_files_from_archive.py (to mimic the dependency of the
   demo towards a "docker_" something).
 - demo_full_workflow.py introduced
 - Demo class:
    * is now abstract
    * knows about an input_demo (for the workflow upstream connectivity)
      (in place of the input_dir)
    * now has a get_output_dir() method
 - LyonMetropoleDowloadAndSanitize renamed to DemoLyonMetropoleDowloadAndSanitize
   and now inherits from Demo. Main section now uses the workflow definitions
 - DemoSplit now correctly places its output in an ouput directory and main
   section uses the workflow definitions (i.e. plugged with an upstream demo)
 - DemoStrip: ditto.
 - docker_extract_building_dates.py now longer imports demo
 - DockerStripAttributes::strip_single_file no longer requires an external
   container.
