+++ Computations/3DTiles/LyonTemporal/PythonCallingDocker/demo_strip_attributes.py
@@ -1,8 +1,9 @@
 import logging
 import os
 import sys
-import docker_strip_attributes
+from docker_strip_attributes import DockerStripAttributes
 from demo import Demo
+import demo_full_workflow as workflow
 
 
 class DemoStrip(Demo):
@@ -13,8 +14,14 @@ class DemoStrip(Demo):
     """
     def __init__(self):
         Demo.__init__(self)
-       
-    def get_result_dir(self, vintage, create=True):
+    
+    @staticmethod
+    def derive_output_file_basename_from_input(input_filename):
+        input_filename = os.path.basename(input_filename)
+        input_no_extension = input_filename.rsplit('.', 1)[0]
+        return input_no_extension + '_stripped.gml'
+
+    def get_vintage_result_dir(self, vintage, create=True):
         """
         :param vintage: a integer or string designating a year
         :param create: when the proposed output directory does not already
@@ -26,21 +33,21 @@ class DemoStrip(Demo):
         if isinstance(vintage, int):
             vintage = str(vintage)
         strip_dir = self.city + '_' + vintage + '_Stripped'
-        result_dir = os.path.join(self.output_dir, strip_dir)
+        result_dir = os.path.join(self.get_output_dir(), strip_dir)
         if create and not os.path.isdir(result_dir):
-            logging.info(f'Creating strip output directory {result_dir}.')
+            logging.info(f'Creating strip output directory {result_dir} '
+                         f'for {vintage} vintage.')
             os.mkdir(result_dir)
         return result_dir
 
-    def get_vintage_borough_resulting_filename(self, vintage, borough):
+    def get_vintage_borough_output_file_basename(self, vintage, borough):
         """
         :return: the filenames (includes the directory name relative
                  to the invocation directory) that the strip algorithm is
                  supposed to produce for the given vintage and borough
         """
-        return os.path.join(
-                self.get_result_dir(vintage, False),
-                borough + '_BATI_' + str(vintage) + '_splited_stripped.gml')
+        input_filename = self.input_demo.get_vintage_borough_output_filename(vintage, borough)
+        return DemoStrip.derive_output_file_basename_from_input(input_filename)
 
     def get_vintage_resulting_filenames(self, vintage):
         """
@@ -50,8 +57,7 @@ class DemoStrip(Demo):
         """
         result = list()
         for borough in self.boroughs:
-            result.append(self.get_vintage_borough_resulting_filename(vintage, 
-                                                                      borough))
+            result.append(self.get_vintage_borough_output_filename(vintage, borough))
         return result
 
     def get_resulting_filenames(self):
@@ -65,35 +71,33 @@ class DemoStrip(Demo):
             result.extend(self.get_vintage_resulting_filenames(vintage))
         return result
 
-    @staticmethod
-    def derive_output_file_basename_from_input(input_filename):
-        input_filename = os.path.basename(input_filename)
-        input_no_extension = input_filename.rsplit('.', 1)[0]
-        return input_no_extension + '_stripped.gml'
-
     def run(self):
+        self.get_output_dir(True)   # Create the output directory
+        input = self.input_demo
         for vintage in self.vintages:
-            vintage_inputs = list()
             for borough in self.boroughs:
-                # FIXME: the name of the output of the split algorithm should
-                # NOT be hardwired here but obtained out of DemoSplit
-                input_filename = os.path.join(
-                     self.output_dir,
-                     borough + '_' + str(vintage),
-                     borough + '_BATI_' + str(vintage) + '_splited.gml')
-
-                docker_strip_attributes.strip_single_file(
-                    docker_strip_attributes.DockerStripAttributes(),
-                    input_dir=os.path.dirname(input_filename),
-                    input_filename=os.path.basename(input_filename),
-                    output_filename=DemoStrip.derive_output_file_basename_from_input(input_filename),
-                    output_dir=self.get_result_dir(vintage))
+                DockerStripAttributes.strip_single_file(
+                    input_dir = \
+                        input.get_vintage_borough_output_directory_name(vintage, borough),
+                    input_filename = \
+                        input.get_vintage_borough_output_file_basename(vintage, borough),
+                    output_filename = \
+                        self.get_vintage_borough_output_file_basename(vintage,borough),
+                    output_dir=self.get_vintage_result_dir(vintage))
 
 if __name__ == '__main__':
-    logger = logging.getLogger(__name__)
-    logging.basicConfig(stream=sys.stdout, level=logging.INFO)
+    strip = workflow.demo_strip
 
-    strip = DemoStrip()
+    logger = logging.getLogger(__name__)
+    logging.basicConfig(level=logging.DEBUG,
+                        format='%(asctime)s %(levelname)-8s %(message)s',
+                        datefmt='%a, %d %b %Y %H:%M:%S',
+                        filename=os.path.join(strip.get_output_dir(True),
+                                              'demo_strip_buildings.log'),
+                        filemode='w')
+    logging.info("Starting to strip files.")
+    
     strip.run()
     strip.assert_output_files_exist()
-    print("Resulting stripped files", strip.get_resulting_filenames())
+    logging.info("Resulting stripped files:")
+    [ logging.info( "   " + file) for file in strip.get_resulting_filenames() ]
